requirements.txt
python-telegram-bot>=13.0
keyring>=23.0
pillow>=8.0.0
watchdog>=2.1.0
